refactor(feature/#75): emotion 컴포넌트 분리
