+++ src/containers/comment/CommentItem.tsx
@@ -1,99 +1,20 @@
 import styled from "@emotion/styled";
-import { useEffect, useState } from "react";
-import { motion } from "framer-motion";
 import { FaCircleUser } from "react-icons/fa6";
-// assets
-import Emotion from "../../assets/emotion.svg";
 // styles
-import { buttonAnimation, iconAnimation } from "../../styles/animation";
 // types
 import { CommentDto } from "../../types/data/comment";
-// api
-import axiosInstance from "../../api/config";
 // hooks
 import useTimeDifference from "../../hooks/useTimeDifference";
+// components
+import Emotion from "./Emotion";
 
 interface CommentItemProps {
   comment: CommentDto;
   fetchComments: () => void;
 }
 
-interface EmotionItem {
-  src: string;
-  alt: string;
-  count: number;
-}
-
-const iconList = [
-  {
-    src: "https://raw.githubusercontent.com/Tarikul-Islam-Anik/Animated-Fluent-Emojis/master/Emojis/Smilies/Grinning%20Face%20with%20Smiling%20Eyes.png",
-    alt: "Smile",
-  },
-  {
-    src: "https://raw.githubusercontent.com/Tarikul-Islam-Anik/Animated-Fluent-Emojis/master/Emojis/Smilies/Red%20Heart.png",
-    alt: "Heart",
-  },
-  {
-    src: "https://raw.githubusercontent.com/Tarikul-Islam-Anik/Animated-Fluent-Emojis/master/Emojis/Smilies/Shaking%20Face.png",
-    alt: "Shock",
-  },
-  {
-    src: "https://raw.githubusercontent.com/Tarikul-Islam-Anik/Animated-Fluent-Emojis/master/Emojis/Smilies/Loudly%20Crying%20Face.png",
-    alt: "Cry",
-  },
-  {
-    src: "https://raw.githubusercontent.com/Tarikul-Islam-Anik/Animated-Fluent-Emojis/master/Emojis/Smilies/Face%20with%20Symbols%20on%20Mouth.png",
-    alt: "Angry",
-  },
-];
-
 const CommentItem = ({ comment, fetchComments }: CommentItemProps) => {
-  const [showEmotions, setShowEmotions] = useState(false);
-  const [selectedEmotion, setSelectedEmotion] = useState<EmotionItem[]>([]);
-
   const timeDifference = useTimeDifference(comment.createAt);
-
-  useEffect(() => {
-    const emotions = Object.entries(comment.reactions)
-      .filter(([_, count]) => count > 0)
-      .map(([key, count]) => {
-        const icon = iconList.find(icon => icon.alt.toUpperCase() === key);
-        return icon ? { ...icon, count } : null;
-      })
-      .filter(icon => icon !== null) as EmotionItem[];
-
-    setSelectedEmotion(emotions);
-  }, [comment.reactions]);
-
-  const showEmotion = () => {
-    setShowEmotions(!showEmotions);
-  };
-
-  const handleEmotionClick = async (
-    commentId: Number,
-    emotion: { src: string; alt: string }
-  ) => {
-    try {
-      const token = localStorage.getItem("token");
-
-      if (!token) {
-        console.log("토큰 없습니다.");
-        return;
-      }
-
-      const url =
-        emotion.alt.toUpperCase() != comment.myReaction
-          ? `/reactions?commentId=${commentId}&emoji=${emotion.alt.toUpperCase()}`
-          : `/reactions?commentId=${commentId}`;
-
-      await axiosInstance.post(url, {});
-
-      fetchComments();
-    } catch (error) {
-      console.log(error);
-    }
-    setShowEmotions(false);
-  };
 
   return (
     <Container key={comment.id}>
@@ -108,47 +29,7 @@ const CommentItem = ({ comment, fetchComments }: CommentItemProps) => {
           <Time>{timeDifference}</Time>
         </Top>
         <Content>{comment.content}</Content>
-        <EmotionContainer>
-          <EmotionBtn
-            onClick={showEmotion}
-            variants={buttonAnimation}
-            initial="hidden"
-            animate={showEmotions ? "visible" : "hidden"}
-          >
-            {showEmotions ? (
-              <IconContainer>
-                {iconList.map(icon => (
-                  <Icon
-                    key={icon.alt}
-                    src={icon.src}
-                    alt={icon.alt}
-                    onClick={() => handleEmotionClick(comment.id, icon)}
-                    variants={iconAnimation}
-                    initial="hidden"
-                    animate="visible"
-                  />
-                ))}
-              </IconContainer>
-            ) : (
-              <Icon src={Emotion} alt="emotion" />
-            )}
-          </EmotionBtn>
-          <SelectedEmotions>
-            {selectedEmotion.map(emotion => (
-              <SelectedEmotionWrapper
-                key={emotion.alt}
-                onClick={() => handleEmotionClick(comment.id, emotion)}
-              >
-                <Icon src={emotion.src} alt={emotion.alt} />
-                <Count
-                  isSelected={comment.myReaction == emotion.alt.toUpperCase()}
-                >
-                  {emotion.count}
-                </Count>
-              </SelectedEmotionWrapper>
-            ))}
-          </SelectedEmotions>
-        </EmotionContainer>
+        <Emotion comment={comment} fetchComments={fetchComments} />
       </InfoWrapper>
     </Container>
   );
@@ -161,7 +42,7 @@ const Container = styled.div`
   grid-template-columns: 30px 1fr;
   grid-gap: 10px;
   margin-bottom: 30px;
-  padding: 0 18px;
+  /* padding: 0 18px; */
 `;
 
 const Image = styled.img`
@@ -197,55 +78,4 @@ const Time = styled.div`
 const Content = styled.div`
   margin-bottom: 5px;
   font-size: ${props => props.theme.typography.paragraphs.default};
-`;
-
-const EmotionContainer = styled.div`
-  display: flex;
-`;
-
-const EmotionBtn = styled(motion.button)`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-right: 5px;
-  border: none;
-  border-radius: 15px;
-  background-color: #363b48;
-  cursor: pointer;
-  overflow: hidden;
-`;
-
-const IconContainer = styled.div`
-  display: flex;
-  justify-content: space-around;
-  width: 100%;
-`;
-
-const SelectedEmotions = styled.div`
-  display: flex;
-`;
-
-const SelectedEmotionWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  margin-right: 5px;
-  padding: 3px 3px;
-  border-radius: 15px;
-  background-color: #363b48;
-`;
-
-const Icon = styled(motion.img)`
-  width: 17px;
-  margin: 0 5px;
-  cursor: pointer;
-`;
-
-const Count = styled.span<{ isSelected?: boolean }>`
-  color: ${props =>
-    props.isSelected ? props.theme.colors.green200 : props.theme.colors.gray50};
-  font-size: ${props =>
-    props.isSelected
-      ? props.theme.typography.disclaimers.bold
-      : props.theme.typography.disclaimers.default};
-  margin: 0 5px 0 2px;
 `;
